Rename shooter-interactive-worker to uicontrol.  Fixes #1889
